+++ gengine/src/systems/LinkSystem.cpp
@@ -1,36 +1,34 @@
 #include <components/Link.hpp>
+#include <components/Position.hpp>
+#include <components/Shape.hpp>
 #include <components/Sprite.hpp>
 #include <components/Text.hpp>
+#include <iostream>
 #include "System.hpp"
 
-void GEngine::System::linkSystem(int id,
-                                 std::map<int, GEngine::Entity>& entities,
-                                 std::pair<float, float> newLinkedEntityPos,
-                                 const int posId) {
+void GEngine::System::linkSystem(int                                id,
+                                 std::map<double, GEngine::Entity>& entities,
+                                 std::pair<float, float>            newLinkedEntityPos,
+                                 const int                          posId) {
     for (auto& [_, entity] : entities) {
-        if (entity.hasComponent<Link>() &&
-            entity.getComponent<Link>().getId() == id) {
+        if (entity.hasComponent<Link>() && entity.getComponent<Link>().getId() == id) {
             GEngine::Entity entityLinked = entities[id];
 
-            auto oldLinkedEntityPos = entityLinked.getComponent<Position>();
-            const std::pair dist = {newLinkedEntityPos.first -
-                                        oldLinkedEntityPos.getPositionX(posId),
-                                    newLinkedEntityPos.second -
-                                        oldLinkedEntityPos.getPositionY(posId)};
+            auto            oldLinkedEntityPos = entityLinked.getComponent<Position>();
+            const std::pair dist               = {
+                newLinkedEntityPos.first - oldLinkedEntityPos.getPositionX(posId),
+                newLinkedEntityPos.second - oldLinkedEntityPos.getPositionY(posId)};
 
-            auto entityPos = entity.getComponent<Position>();
-            const std::pair newPos = {entityPos.getPositionX(0) + dist.first,
-                                      entityPos.getPositionY(0) + dist.second};
+            auto            entityPos = entity.getComponent<Position>();
+            const std::pair newPos    = {
+                entityPos.getPositionX(0) + dist.first, entityPos.getPositionY(0) + dist.second};
 
             if (entity.hasComponent<Sprite>()) {
-                updatePosition(entity,
-                               entity.getComponent<Sprite>().getSprite(),
-                               newPos, 0);
+                updatePosition(entity, entity.getComponent<Sprite>().getSprite(), newPos, 0);
             }
 
             if (entity.hasComponent<Text>()) {
-                updatePosition(entity, entity.getComponent<Text>().getText(),
-                               newPos, 0);
+                updatePosition(entity, entity.getComponent<Text>().getText(), newPos, 0);
             }
         }
     }
